:tractor: relocate copier to plugins
